Here you go Chase more comments
Waluigi is the ultimate example of the individual shaped by the
signifier. Waluigi is a man seen only in mirror images; lost in a hall
of mirrors he is a reflection of a reflection of a reflection. You
start with Mario – the wholesome all Italian plumbing superman, you
reflect him to create Luigi – the same thing but slightly less. You
invert Mario to create Wario – Mario turned septic and libertarian –
then you reflect the inversion in the reflection: you create a being
who can only exist in reference to others. Waluigi is the true nowhere
man, without the other characters he reflects, inverts and parodies he
has no reason to exist. Waluigi’s identity only comes from what and who
he isn’t – without a wider frame of reference he is nothing. He is not
his own man. In a world where our identities are shaped by our warped
relationships to brands and commerce we are all Waluigi.